+++ lib/groupdate/magic.rb
@@ -158,7 +158,8 @@ module Groupdate
 
     def time_zone
       @time_zone ||= begin
-        time_zone = options[:time_zone] || Groupdate.time_zone || Time.zone || "Etc/UTC"
+        time_zone = "Etc/UTC" if options[:time_zone] == false
+        time_zone ||= options[:time_zone] || Groupdate.time_zone || Time.zone || "Etc/UTC"
         time_zone.is_a?(ActiveSupport::TimeZone) ? time_zone : ActiveSupport::TimeZone[time_zone]
       end
     end